Circle: number of segments is power of 2
